+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/commands.py
@@ -107,7 +107,7 @@ def load_command_table(self, _):
         g.custom_command('set', 'update_connection_strings')
         g.custom_command('delete', 'delete_connection_strings')
 
-    with self.command_group('webapp config azure-storage-accounts') as g:
+    with self.command_group('webapp config storage-account') as g:
         g.custom_command('list', 'get_azure_storage_accounts', exception_handler=empty_on_404)
         g.custom_command('add', 'add_azure_storage_account')
         g.custom_command('update', 'update_azure_storage_account')