+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 +++++
 * appservice plan, webapp & function app updated to use latest python websites SDK version
 * az webapp config set supports --fts-state property. Also added support fot az functionapp config set & show
+* webapp: add support for bring your own storage
 
 0.2.3
 +++++